Added new button to addEventMenu
For testing, added an extra button to test addEvent() function from addEvent.js file. If functional, Add Event 2 will create a new event and add it to the database, if it is legal. Cannot implement radio checkboxes at this time, due to difficulty involving getting information about which option is selected at the time of button press.